Refactor API groups extension adapter to make it easier for future API groups
